post changed to http.post (#19)
* TODO message on auth_model.dart

* can check if restriction is empty

* Routes changed in initial page

* functional autologin and logout

* loading and snackbar added for logout

* printed errors removed

* restriction nullable in model, initial page and login page

* adminPage and landingPage are already working

* SDK upgrade

* in authremote, post is changed to http.post

---------

+++ lib/features/auth/data/datasource/auth.local.datasource.dart
@@ -46,10 +46,21 @@ class AuthlocalDatasource {
     }
   }
 
+  Future<Unit> containsKey() async {
+    try {
+      if (await _secureStorage.containsKey(key: Config.userToken)) {
+        await _secureStorage.delete(key: Config.userToken);
+      }
+      return unit;
+    } catch (e) {
+      throw e.toString();
+    }
+  }
+
 //for logout
   Future<Unit> deleteToken() async {
     try {
-      _secureStorage.delete(key: Config.userToken);
+      await _secureStorage.delete(key: Config.userToken);
       return unit;
     } catch (e) {
       throw (e.toString());
@@ -58,7 +69,7 @@ class AuthlocalDatasource {
 
   Future<Unit> deleteUser() async {
     try {
-      _secureStorage.delete(key: Config.user);
+      await _secureStorage.delete(key: Config.user);
       return unit;
     } catch (e) {
       throw (e.toString());
